Update office app name
Hi :)

I was following the instructions and was confused that it told me to install a "Collabora Online" app, because if you search for this in the appstore, the result is [Collabora Online - Built-in CODE Server](https://apps.nextcloud.com/apps/richdocumentscode). If I understand correctly, it should be [Nextcloud Office](https://apps.nextcloud.com/apps/richdocuments).

It seems as if Nextcloud Office has changed its name multiple times, so I added the link to the store as well.

Fixed some layout while I was at it; hope that's okay.

Best regards!

Signed-off-by: Dennis Müller <github.dm@discovery-one.de>